# solution/backend/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
...
paths:
  /api/auth/setup-password:
    post:
      tags:
      - Auth
      summary: Set password after first login (when requiresPasswordSetup=true)
      operationId: setupPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupPasswordRequest"
        required: true
      responses:
        "400":
          description: Passwords do not match
        "401":
          description: Not authenticated
        "204":
          description: Password set
  /api/auth/resend-activation:
    post:
      tags:
      - Auth
      summary: Resend activation link to a PENDING account (US-REG-03)
      operationId: resendActivation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendActivationRequest"
        required: true
      responses:
        "204":
          description: Request processed — email sent if account is PENDING
        "400":
          description: Invalid email format
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Buyer self-registration (FS-B01) — sends activation email
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Account created — activation email sent
        "409":
          description: Email already registered
        "400":
          description: Invalid input
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Login — returns a signed JWT
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "401":
          description: Invalid credentials or account not active
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "200":
          description: Authentication successful
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/activate:
    post:
      tags:
      - Auth
      summary: Activate account via email token (US-ADM-01 / US-REG-02)
      operationId: activate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateAccountRequest"
        required: true
      responses:
        "204":
          description: Account activated
        "410":
          description: Token expired or not found
        "400":
          description: Password required or mismatch
...
components:
  schemas:
    SetupPasswordRequest:
      required:
      - confirmPassword
      - password
      type: object
      properties:
        password:
          maxLength: 2147483647
          minLength: 12
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&_\\-#]).{12,}$"
          type: string
          description: "Chosen password — minimum 12 characters, must contain uppercase,\
            \ lowercase, digit and special character"
        confirmPassword:
          type: string
          description: Password confirmation — must match password
    ResendActivationRequest:
      required:
      - email
      type: object
      properties:
        email:
          type: string
          description: Email address of the PENDING account requesting a new activation
            link
    RegisterRequest:
      required:
      - email
      - firstName
      - lastName
      - password
      type: object
      properties:
        email:
          type: string
          description: "Email address, unique across the platform"
        password:
          maxLength: 2147483647
          minLength: 12
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&_\\-#]).{12,}$"
          type: string
          description: "Chosen password — minimum 12 characters, must contain uppercase,\
            \ lowercase, digit and special character"
        firstName:
          type: string
          description: First name
        lastName:
          type: string
          description: Last name
    LoginRequest:
      required:
      - email
      type: object
      properties:
        email:
          type: string
          description: Admin account email address
        password:
          type: string
          description: Account password — omit on first login
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Signed JWT to include in subsequent requests as Bearer token
        email:
          type: string
          description: Authenticated account email
        requiresPasswordSetup:
          type: boolean
          description: True when the account has no password yet — client must call
            /api/auth/setup-password
    ActivateAccountRequest:
      required:
      - token
      type: object
      properties:
        token:
          type: string
          description: Activation token received by email
        password:
          maxLength: 2147483647
          minLength: 12
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&_\\-#]).{12,}$"
          type: string
          description: "Chosen password (required when account has no password yet)\
            \ — minimum 12 characters, must contain uppercase, lowercase, digit and\
            \ special character"
        confirmPassword:
          type: string
          description: Password confirmation
...